backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
pytesseract==0.3.13
Pillow==10.4.0
python-dotenv==1.0.0
spacy==3.6.0
gunicorn==20.1.0
stripe==5.0.0
Flask-SocketIO==6.5.2
eventlet==0.33.3
APScheduler==3.10.1
